requirements.txt
# Requirements for Enhanced Structured RAG System

# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.8

# LangGraph for ReAct agents
langgraph>=0.0.20

# Vector Database
faiss-cpu>=1.7.4

# OpenAI
openai>=1.12.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Environment and Configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Data Processing
pydantic>=2.5.0

# Async support
asyncio
aiohttp>=3.9.0
nest-asyncio>=1.5.6  # For nested event loop support

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0  # Optional for data analysis

# Development tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0